requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
datasets>=2.12.0
Pillow>=9.5.0
opencv-python>=4.8.0
faiss-cpu>=1.7.4
streamlit>=1.25.0
python-dotenv>=1.0.0
